PLT-7050 Included timeline in the result report view
